Introduce session type parameter to extension features
This adds session type to extension features grammar. The session
type refers to the type of user active in the session in which
an extension is running. The parameter will enable making a feature
available depending on type of the session in which the extension/app
is running.

Feature session types are declared as list of individual session types
in which a feature is available (undefined list means no restrictions).
Currently, values supported for session type are as follows:
* 'kiosk' - kiosk app session
* 'regular' -  non-kiosk user session
The list of supported values is expected to be expanded in future.

Depends on https://codereview.chromium.org/2241203003/ for plumbing
of current session type to feature availability checks

BUG=606417

Review-Url: https://codereview.chromium.org/2255613003
Cr-Original-Commit-Position: refs/heads/master@{#418938}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64bf38deace6fc500ab416cae16f9d3ea0629411