chromeos: Add a flag to launch with --mash.
Add a flag in chrome:flags to turn on the mojo ui service. The caveat
is that it works OK (as in, you can successfully login) if a secondary
user on the device turns on the flag. If the primary user/owner turns
the flag on, then the flag is used in the login screen as well, and
that does not work well yet.

Even though secondary/non-owner users can turn the flag on and login,
and see the chrome browser window (and other mus apps) without getting
immediate crashes, a number of things are still in progress, and not
implemented yet. So there are a number of missing features, and is
somewhat unstable.

BUG=633656

Review-Url: https://codereview.chromium.org/2359483003
Cr-Original-Commit-Position: refs/heads/master@{#420219}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 59e835de462f99529c7312b49926e2f0a67aa36e