Add integration tests for QUIC
Two tests have been added: First is a smoke test (only cares
about the end result -- if the page was loaded or not etc.),
second looks more closely to verify if QUIC was actually used
or not.

Also, removed the global --disable-quic flag, and instead
added it to the individual tests that needed it.

Verified that all benchmarks currently pass on Android.

BUG=343579

Review-Url: https://codereview.chromium.org/2357723002
Cr-Original-Commit-Position: refs/heads/master@{#420390}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 285e6f35b3b135cb2975a47afb59bed5087225d6