[CustomTabsBenchmark] Support prerendering mode
This CL allows to set the prerendering mode to disabled,
enabled, or prefetch.

This replaces the old 'no_prerendering' boolean.

The prerendering mode is passed to the java application
(changing its type from a boolean to a string), and to
Chrome as a command line switch.

Review-Url: https://codereview.chromium.org/2372193002
Cr-Original-Commit-Position: refs/heads/master@{#421487}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51d7f96fac92e6c0707fb49b0476beb37c0ddde4