Record download target connection security state
As part of HTTP-bad effort, we want to learn about and then indicate the
insecure state of HTTP downloads. This CL add UMA metric to record the
connection security state of download target, indicating whether the final
resolved download url and the redirects before it are secure or not.

BUG=645521

Review-Url: https://codereview.chromium.org/2378333002
Cr-Original-Commit-Position: refs/heads/master@{#422148}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82b8e5071d4088fc18daf80f6f49047ef5a6ff5e