bluetooth: Create RequestDevice.UnionOfServices.Services UMA
UMA of required and optional services exist, but are difficult to
interpret results because applications may be using different mixes
of required / optional in their filters.

This new UMA for the union provides a single location to measure
per-connection intent to intaract with UUIDs.

Review-Url: https://codereview.chromium.org/2367173002
Cr-Original-Commit-Position: refs/heads/master@{#422159}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2eff8ec36aa5ad705b2997cee07fd5a1436336b1