predictors: Take the priority into account for scoring resources.
This changes the way the resource score is computed in the
resource_prefetch_predictor. It takes into account the request priority
as the most important feature, then resource type, then average position.

The rationale is that resource priority surfaces the render-blocking
resources, that are the most critical to prefetch.

BUG=631966

Review-Url: https://codereview.chromium.org/2385173002
Cr-Original-Commit-Position: refs/heads/master@{#422753}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3eb631ab216ff9ea0302a9e89efe33003b11d935