Removed Ash.AppList.TimeBetweenTaskSwitches and Ash.Tab.TimeBetweenSwitchToExistingTabUserActions histograms.
These histograms are triggered in the chrome process but are actually
recorded in ash (which is in a different process). This is problematic
for the mus+ash refactoring and since there are no clear decisions
being made from these metrics they are being removed.

BUG=616581
TEST=ash_unittests

Review-Url: https://codereview.chromium.org/2390793002
Cr-Original-Commit-Position: refs/heads/master@{#422885}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a7916b557200bb2ea8291732158e200f434078b