customtabs: Delay logging of TTFCP in the benchmark app.
The Time To First Contentful Paint (TTFCP) benchmark can be dispatched
after the NAVIGATION_FINISHED event. In this case, wait up to 3s for it
to come before logging the result.

Also increase the delay in customtabs_benchmark.py, as it is limiting on
low-end devices.

Based on a patch by droger@.

Review-Url: https://codereview.chromium.org/2397753002
Cr-Original-Commit-Position: refs/heads/master@{#423137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ffb1ebb3a7a1486984f2b2a0133c0ac20d0dadb1