Simplify the text in the external protocol confirmation dialog.
The external protocol handler dialog currently presents users with
a very large amount of jargon-filled text. This CL simplifies the dialog
down to just a title containing the app name and the existing checkbox.

On Windows 7 and earlier, the registry is queried for an app path,
which is displayed in the dialog. This CL modifies the query to
retrieve a program name from the app path, falling back to just
the executable name if the retrieval doesn't succeed.

A new histogram to log the number of times users check the
checkbox in the dialog is also added.

BUG=601725

Review-Url: https://codereview.chromium.org/2076253002
Cr-Original-Commit-Position: refs/heads/master@{#423664}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0c9a5065a8a7c4b760b79eb996548b1921ed05ea