UMA histogram for longest renderer main thread usage for page save to MHTML.
The information about the time taken by the main render process thread
for saving a page to MHTML is the main feedback we have to see how much
the operation impacts user experience.

The existing UMA that tracks this information,
PageSerialization.MhtmlGeneration.RendererMainThreadTime.SingleFrame,
is not directly useful for the purpose above because it includes reports
for all frames in a page save operation. For pages with many frames, most
of them are very simple and quickly saved, while one -- generally the
main frame -- is the slowest. So we have a huge amount of low-value
samples pulling the statistics down.

This change adds a new histogram,
PageSerialization.MhtmlGeneration.RendererMainThreadTime.SlowestFrame,
that reports only the longest of all times from the former one in a
single HTML page saving operation, which is the most important one in
terms of user experience.

BUG=645686

Review-Url: https://codereview.chromium.org/2400383002
Cr-Original-Commit-Position: refs/heads/master@{#424312}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90bf4f0c77709ddb65b1f8d86e04065aacae02ac