customtabs: Add a flag to mayLaunchUrl() to force prefetch.
This adds a flag to mayLaunchUrl() to prevent it from doing a prerender,
forcing the high confidence version to only issue a prefetch
request (using the resource_prefetch_predictor here). This is needed to
test and evaluate it, and is intended for testing purposes only, as the
"no prerendering" flag is as well.

BUG=650250,631966

Review-Url: https://codereview.chromium.org/2403023002
Cr-Original-Commit-Position: refs/heads/master@{#424742}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 588e88df5847e11703ffdf063f621d38a66a5fa4