Support subdomain matching in trial tokens
This CL adds support for subdomain trial tokens, such that a properly
issued token will enable a trial for multiple subdomains, rather than a
single specified origin.

In bug 653349, it was proposed to use a leading "*" to indicate a
wildcard token (e.g. *.example.com). In this implementation, an explicit
flag was added to the token format. The explicit flag keeping the token
parsing simple, by keeping the url::Origin as the representation of the
origin, and avoiding custom string parsing.

BUG=653349

Review-Url: https://codereview.chromium.org/2411803002
Cr-Original-Commit-Position: refs/heads/master@{#425168}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f0cb8e9b16a270d0a7fd313f161b4c2c9a0df29