[NTP Snippets] Report the share of the UI lang vs. the top non-UI lang
In requests to content suggestions server, we send the frequency of use
of the UI language and the top non-UI language.

This CL introduces reporting of the ratio of these two frequencies.
This allows to see how often (and how much) the UI language is
inappropriate.

BUG=656011

Review-Url: https://codereview.chromium.org/2419013003
Cr-Original-Commit-Position: refs/heads/master@{#425656}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: caa3c1cbea3a8de6603ad2f1e59d14224f4167f1