Linux build: Use sysroot when calculating dependencies
Additionally, this CL also bundles part of dpkg-shlibdeps.  The version of dpkg-shlibdeps was too old on Precise to support the feature we needed for using the sysroot to calculate dependencies (specifically, the "-l" option was needed).  This meant we needed to temporarily bundle a newer version dpkg-dev to get this working on Precise.

Trusty has a newer version of dpkg-dev with the necessary feature, so dpkg-dev should be removed once the builders are entirely switch off of Precise.

This can be done by reverting this CL and applying ps40001 from this CL.

BUG=564904
R=thestig@chromium.org,sbc@chromium.org,jochen@chromium.org

Review-Url: https://codereview.chromium.org/2411423002
Cr-Original-Commit-Position: refs/heads/master@{#425842}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0db721ca35ac387f5adf6bb9d7ec5ec24be4ac6e