cc: Add back a benchmark category trace for counting renderer frames
When deleting DelegatingRenderer::SwapBuffers I failed to notice that
the TRACE_EVENT in there had the benchmark category. This category
is used by perf tests, and without that event they had no way to count
the number of renderer frames produced.

This adds benchmark to the DrawLayers trace event in LayerTreeHostImpl
instead and points the perf tests at that. Also removes the benchmark
category from the GLRenderer and SoftwareRenderer classes as they are
never used in the renderer and are not used for perf tests (tools/ no
longer refers to the name "SwapBuffers" for perf tests).

R=enne@chromium.org
TBR=dtu
BUG=656947
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_precise_blink_rel

Review-Url: https://chromiumcodereview.appspot.com/2442453003
Cr-Original-Commit-Position: refs/heads/master@{#426952}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7607802c76a4a0b826d22807213e981025ec734b