[Sync] Start implementation of migration for USS.
This approach draws data and metadata from the server half of the
directory and injects it into USS as the initial GetUpdates response.

This CL does not actually enable migration in production because
an empty migrator function is passed instead of the real one.

BUG=658002

Review-Url: https://codereview.chromium.org/2442583003
Cr-Original-Commit-Position: refs/heads/master@{#428467}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 822cfab0275712dbd7d69673f68a32461b53c1bc