Increase peeking card visibility.
The previous peeking card would show as the top of a card peeking above the bottom of the screen, expanding to full width as the user scrolled. This showed no content.

This changes the peeking card to be full width and to animate (based on time) up from the bottom of the screen, coming to a resting place that shows the first half line of the title.

BUG=660837

Review-Url: https://codereview.chromium.org/2468973004
Cr-Original-Commit-Position: refs/heads/master@{#429962}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9883eff13a648ff47c37d912cc46a2cc30191030