[Re-land] Add network throttling as enterprise policy
This is a re-land of codereview.chromium.org/2364703002/
which was reverted in https://codereview.chromium.org/2461383002/
due to a memory leak in network_throttling_observer_unittest causing
the long-running linux-chromeos ASAN bot to fail.
All files stay the same except for the policy enum advancing to 351 from 350 (350 was taken while this CL was reverted) and memory leak fixes to network_throttling_observer_unittest.cc

Add support for an enterprise policy that lets the admin
turn on network throtttling and specify the upload/download
rates in kbps to throttle to.

BUG=642811
TBR=atwilson@chromium.org,brettw@chromium.org,stevenjb@chromium.org
TEST=(1) unit-tests (2) Added network policy to YAPS
server, enabled throttling with various values, disabled
throttling and checked that the values propagate to client

Review-Url: https://codereview.chromium.org/2466693002
Cr-Original-Commit-Position: refs/heads/master@{#430082}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 857eeac9ace6b79bd8b99fceac315455ee82f617