Update the Google Search geolocation disclosure to make it more obvious.
This Android change (behind a flag) is the first in a series of patches
to make the Google Search geolocation disclosure more noticeable.

When the flag is enabled, this patch will:
- disable the existing disclosure
- show a new infobar disclosure when
  - an omnibox search has just been done
  - Google search is the default search engine
  - Chrome has geolocation permission
  - The current default search origin has not had geolocation allowed
    or blocked by the user.

Note, this patch shows the disclosure on every navigation that meets
these conditions.

Future patches will limit how many times the disclosure is shown, will
update the infobar to have a link, and update the infobar to always
be on top of other infobars.

BUG=661011

Review-Url: https://codereview.chromium.org/2475213002
Cr-Original-Commit-Position: refs/heads/master@{#430516}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3503a8fc3531a8b903aef6db34bed83a73d2960d