Add a PRESUBMIT check for broken .ico files.
This is in reaction to crbug.com/526622, where the Chrome .ico files
had a bad mask, making them render incorrectly on Windows. The presubmit
now checks that icons have a properly formed mask, as well as that the
large image is a compressed PNG (not an uncompressed BMP).

Also adds a --lint argument to the optimize-ico-files script that
manually runs the same checks on a given .ico file.

BUG=534679

Review-Url: https://codereview.chromium.org/1372113003
Cr-Original-Commit-Position: refs/heads/master@{#430777}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e56b0cf9ec1e78b04c254405643c7a74380be168