Remove the GN cookbook document.
This document was for listing equivalents between GYP and GN which are no longer relevant.

Many of the items our out-of-date already, and it is not being updated further. I have already removed the references to this document from the Google Sites pages.

Review-Url: https://codereview.chromium.org/2505133004
Cr-Original-Commit-Position: refs/heads/master@{#432897}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6afdb84319431eb94809a5f2ef9f33960dd2dea6