Convert one more CFI blacklist entry into DISABLE_CFI_PERF attribute.
This one is tricky as it broke Chrome OS build last time. The line
now moved below the template parameter, and that should work.

The primary problem here is that Chrome OS lacks a trybot here,
so it's impossible to know in advance, if the change is going to break it.

BUG=641536
TBR=eae

Review-Url: https://codereview.chromium.org/2517993003
Cr-Original-Commit-Position: refs/heads/master@{#433684}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 565ff1e9c4a5bcda8db37bdd8365a7eee9f3c029