cygprofile: workaround: look for cyglog files deeper
On the bot self._device.PullFile() returns ['/.../cyglog_data/cyglog'], which
points to a directory with files (that we need). In my case the list contains
the files under this directory, which is the intended behavior.

I cannot reproduce the behavior on the bot even though the relative path to the
'adb' executable is the same on my desktop. Hence, a (hopefully temporary)
workaround.

BUG=645460

Review-Url: https://codereview.chromium.org/2520353005
Cr-Original-Commit-Position: refs/heads/master@{#434189}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 509e990fe476c406a7d9510df9fa8ccf7c8d0f97