Add sub-metrics for Event.Latency.TouchToFirstScrollUpdateSwapBegin.
We defined a set of sub-metrics to measure the latencies inside
Event.Latency.TouchToScrollUpdateSwapBegin to help us better understand
the time spent on each stage during scroll. Now we will add the same set
of sub-metrics to Event.Latency.TouchToFirstScrollUpdateSwapBegin.

The details are in this doc:
https://docs.google.com/document/d/1EPKT652e2PcO1ysA7LQosGyyr1xmxbtiiiivNU3lnpQ/edit#heading=h.r6af7ldx7kcw.

BUG=616482

Review-Url: https://codereview.chromium.org/2268163002
Cr-Original-Commit-Position: refs/heads/master@{#434682}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb43b09f06395377e66c684522d76ec002563fb0