Add a timeout to push message processing.
Currently, the only time limitation on push message processing is the
service worker default timeout of 5 minutes. 99% of push messages are
processed in under 1 minute, so we want to reduce the timeout so that
outliers don't begin to rely on the long processing available.

BUG=646783

Review-Url: https://codereview.chromium.org/2506393003
Cr-Original-Commit-Position: refs/heads/master@{#434995}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0046294f2f1c6a3b72dd6d373a8530e1a7a296ec