Support API aliases
Introduce API feature properties to specify an API source and alias.

|source| is set on API alias, and is used during API registration
- if an API feature has source set, the bindings installed under API
name will be the bindings of |source| API.

|alias| is used on source API, and is used when checking API
availability - when checking availability of an API that has alias
set, the API will be deemed available if either one of the API and
the API alias are available in the context.
Note that this logic should not be used when determining API
availability for binding registration - in this case alias
availability should be ignored in order to avoid registering
source API bindings when only alias bindings are needed.

BUG=647415

Review-Url: https://codereview.chromium.org/2494653005
Cr-Original-Commit-Position: refs/heads/master@{#435034}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: feb4b052b4c86b9047f8c220e3ec3983e5483cdf