tools: Local tests for the speculative prefetch predictor.
Adds a tool to reset Chrome's state, push a predictor database to a
device, and load a page after a prefetch, collecting loading metrics.
This is a follow-up to crrev.com/2496563002.

Also adds a way to only enable external prefetch requests in the predictor.

BUG=655980

Review-Url: https://codereview.chromium.org/2508933002
Cr-Original-Commit-Position: refs/heads/master@{#435270}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b65cb1a116e559085ec372ea2ea4891bd00b1181