Script for parsing of verbose linker output
When trying to understand why particular symbols are pulled in to a
binary it can be helpful to examine the verbose linker output.
However it quickly becomes unwieldy. This parses it and allows
querying of it to explain why a particular object file is pulled
in. This was used in the investigation of the linked bug, and others.

BUG=630755

Review-Url: https://codereview.chromium.org/2564543003
Cr-Original-Commit-Position: refs/heads/master@{#437730}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de77b2055e40cab9321479335ffd3b8416c3e34c