Introduces alias networking.onc for networkingPrivate API.
Adds networking.onc permission, which guards the introduced
networking.onc alias - for now, the permission is bound to dev channel
and available to the same whitelist as networkingPrivate API.

To avoid duplicating whitelist for the two permissions, the whitelist
is moved to behavior features and the behavior feature is added as a
dependency for both of them.

Refactors networking private API tests to enable parameterizing the
networking API to be used in the test and adds tests that exercise
alias - networking.onc code path.
Also groups some sub tests (which were all run in a separate browser
test) together in order to reduce total overhead of starting browser
for browser tests.

This is the first step towards renaming networkingPrivate API to
networking.onc (which will be made available to kiosk sessions).

BUG=672186

Review-Url: https://codereview.chromium.org/2470193002
Cr-Original-Commit-Position: refs/heads/master@{#438574}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 975d838eca97d3cac1578861c9df42fe46e2a9d5