[Chrome OS] Add a chrome os device wallpaper policy.
Today, even if domain admins set a wallpaper policy, every device
shows up with the branding of the OEM until a user logs into the
device (and then the user wallpaper policy kicks in). We introduce
a new device policy in this CL for domain admins to set a device
level policy rather than a user level policy for setting a wallpaper.
That way the wallpaper will change after being enrolled to the domain
without requiring users to log in. This only impacts the startscreen
wallpaper, if the domain admin has set a different wallpaper in the
user policy or has allowed the users to select their own wallpaper,
that one will be displayed within a session.

Note this CL only introduces the policy, the implementation to reflect
the policy is addressed in another CL:
https://codereview.chromium.org/2572793003/.

BUG=475998

Review-Url: https://codereview.chromium.org/2544693002
Cr-Original-Commit-Position: refs/heads/master@{#438592}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0823eab840a0a2c22692be4c21870a0c51e07e0e